+++ test/dune/eval/opts_test.exs
@@ -1,4 +1,0 @@
-defmodule Dune.Eval.OptsTest do
-  use ExUnit.Case, async: true
-  doctest Dune.Eval.Opts
-end