+++ test/dune/opts_test.exs
@@ -1,0 +1,4 @@
+defmodule Dune.OptsTest do
+  use ExUnit.Case, async: true
+  doctest Dune.Opts
+end