+++ test/dune/parser/opts_test.exs
@@ -1,4 +1,0 @@
-defmodule Dune.Parser.OptsTest do
-  use ExUnit.Case, async: true
-  doctest Dune.Parser.Opts
-end